requirements.txt
streamlit
pandas
scikit-learn
scipy
numpy
scikit-surprise
